:bug: fixed can not create hybrid provider instance
